refactor: `model`という一つのパッケージに様々な意味合いのクラスが集まっており、分かりにくくなっていたため、役割に応じてパッケージを分ける。詳細は後述
- `bean`というパッケージを追加。リクエストやレスポンスなど、主にController層で扱われる情報が配置される
- `dto`というパッケージを追加。主にService層で扱われる、ビジネスロジックに関する情報が配置される
- `entity`というパッケージを追加。主にDBのテーブルなど、主にRepository層で扱われる情報が配置される
- `model`というパッケージは削除
- パッケージの命名は、前職で使っていたアーキテクチャを踏襲している。問題がありそうであれば、別途修正する。

refs #8 出費情報を扱うモデルの整理
