+++ src/main/kotlin/com/thurofuji/expensesBook/bean/ExpenseResponse.kt
@@ -1,4 +1,4 @@
-package com.thurofuji.expensesBook.model
+package com.thurofuji.expensesBook.bean
 
 import java.time.LocalDate
 import java.util.UUID
