+++ src/main/kotlin/com/thurofuji/expensesBook/dto/ExpenseDto.kt
@@ -1,4 +1,4 @@
-package com.thurofuji.expensesBook.model
+package com.thurofuji.expensesBook.dto
 
 import java.time.LocalDate
 import java.util.UUID
