+++ src/main/kotlin/com/thurofuji/expensesBook/dxo/ExpenseBookDxo.kt
@@ -1,9 +1,9 @@
 package com.thurofuji.expensesBook.dxo
 
-import com.thurofuji.expensesBook.model.ExpenseDto
-import com.thurofuji.expensesBook.model.ExpenseType
-import com.thurofuji.expensesBook.model.NewExpenseDto
-import com.thurofuji.expensesBook.model.出費履歴
+import com.thurofuji.expensesBook.dto.ExpenseDto
+import com.thurofuji.expensesBook.dto.ExpenseType
+import com.thurofuji.expensesBook.dto.NewExpenseDto
+import com.thurofuji.expensesBook.entity.出費履歴
 import java.util.UUID
 
 /**
